Refac: Pass all constants files to contants file Add: New file with commom project constants
