+++ src/lime/protocol/__init__.py
@@ -1,9 +1,7 @@
 from .serializable import *
 from .envelope import *
-from .command_method import *
-from .command_status import *
 from .command import *
 from .node import *
 from .identity import *
-from .notification_event import *
 from .reason import *
+from .constants import *
