+++ src/lime/protocol/constants/common_constants.py
@@ -1,0 +1,7 @@
+class CommonConstants:
+    """Commom constants used in project."""
+
+    DOMAIN = 'take.net'
+    NAME = 'take'
+    DOMAIN_SEPARATOR = '@'
+    INSTANCE_SEPARATOR = '/'
