+++ src/lime/protocol/constants/notification_event.py
