+++ src/lime/protocol/constants/session_compression.py
