--- tests/protocol/test_contants.py
@@ -1,7 +1,0 @@
-class TestConstants:
-    """Commom constants used in tests."""
-
-    DOMAIN = 'take.net'
-    NAME = 'take'
-    DOMAIN_SEPARATOR = '@'
-    INSTANCE_SEPARATOR = '@'
